Added nidhins translate functions

+++ Group17RobotReal/Group17Robot.cydsn/motor.h
@@ -12,7 +12,7 @@
 
 // Functions: 
 void moveDynamic(int distance, int speed, int activate_safety);    // positive number to move forward by that distance, negative to move backwards
-void translateMoveDynamic(int distance, int degree, int speed, int activate_safety);    // positive number to move right by that distance, negative to move left
+void translateMoveDynamic(float distance, int degree, int speed, int activate_safety);    // positive number to move right by that distance, negative to move left
 void moveSwivel(int degrees, int speed, int activate_safety);
 void rampUpDown(int speed, int ramp_up_down);
